Initial checkin of the distributed forks tooling.
This repository implements distributed tracking and management of git
forks by storing the list of forks in the repository itself.

The idea behind this was first proposed as a feature of
[`git-appraise`](https://github.com/google/git-appraise/pull/88), but
is being split out into a standalone feature after it became clear
that the functionality was useful in more contexts than just for code
review.