Simplify the default fetchspecs for forks.
In the new model, heads, tags, and notes are treated specially,
but all refs from a fork are also mapped under `refs/forks/<name>/`,
so that no refs are missed.

Since this means that a single ref from a fork will become multiple
refs in the virtual repository, this change also had to extend the
logic in the remote helper to support this.